[ar] Allow a plugin resolver to declare itself as primary
A plugin resolver could only be considered a primary resolver if it does not specify uriSchemes in plugInfo.json. This prevents a "fallback" sort of behavior when multiple plugin resolvers that support URI schemes need to live in the same environment but one is intended to be the primary resolver

This change adds a "primary" flag for a plugin resolver in plugInfo.json to explicitly set. The "primary" flag indicates that the plugin resolver could be a primary resolver but also supports specific URI schemes