[hd, hdSt, usdImaging] Enables storm to get the correct identifier when using scene indices.
The previous sceneDelegate has the render context and is able to use
that to convert the "sourceAsset" into a node identifier.  With the new
scene index, we instead read all the relevant information ("info:*" and
some metadata) and store that in the renderContextNodeIdentifiers
container.  The HdMaterialNetworkInterface now has functions for
accessing that data.  This allows us to make a matfilter
(nodeIdentifierResolvingSceneIndex) that reads the data at "info:*" and
then set the node type ("nodeIdentifier").

(Internal change: 2247027)