Run HdsiTerminalsResolvingSceneIndex earlier.
Scene Index filters that need to access/changematerials should be able to do it
in a renderer agnostic way. This means that the terminal names should already be
resolved (with no renderer prefix) at that point, which can be early in the scene
index stack. One example that expects this is _MaterialLayerSetupSceneIndex,
which was running before the terminal resolution.

(Internal change: 2248848)