[hdx, usdAppUtils, usdImagingGL] Enable usdrecord to disable the GPU when generating images.
This mode is enabled by passing the "--disableGpu" option to the usdrecord.py script.  The GPU is enabled by default.  If no renderer is specified, an appropriate default will be chosen depending on whether the GPU is enabled or not.

Additionally, disabling the GPU will prevent the HdxTaskController from creating any tasks that require it.  In particular, this means that color correction is disabled when the GPU is disabled.  A warning to this effect will be emitted if color correction is requested in this case.

Fixes #1926

(Internal change: 2251239)
(Internal change: 2251338)