Add support for opaque-typed attributes in USD.
This adds a new typename, "opaque", which is used for attributes that do
not have a value that is representable in USD. Opaque attributes are
intended for applications that want to have attributes that can be
connected to other attributes, or can be targeted by relationships,
indicate dataflow or other dependency information, without that attribute
having a value that can be represented in a USD layer.

Opaque attributes are just like any other attribute except that their
values cannot be persisted to (or loaded from) a layer; opaque attributes
are defined to have values that are SdfOpaqueValue typed, and
SdfOpaqueValue is an empty struct. Other than the special handling of their
values, though, they are just like any other attribute; any other field can
be authored on them normally, and they can be targeted by connections and
relationships.

This also adds another new typename, "group", which is defined to be
opaque-valued with role "group". Group attributes are intended to be used
to provide a connectable and targetable object that represents a property
namespace; for example, an attribute "a:b:x" would be a member of group
"a:b", if a group attribute named "a:b" existed.

Since USD has always supported loading attributes with unknown typenames,
this does not constitute a new scene description version; old versions of
USD will continue to load new files just fine. In fact, the only
perceptable difference is that some files that used to load will stop
loading. When we see an unrecognized typename, USD surfaces the authored
value of that attribute as an "unrecognized value", with a string
representation of what was in the file, so previously it was valid to have
a line like "opaque Attr = 1.0". That same file now will fail to parse,
because "opaque" is a recognized typename and the parser throws an error
any time it tries to deserialize an authored value of type SdfOpaqueValue.
In order to ensure this is the case, both the cratefile and text formats
are updated to ensure that they will not write out an SdfOpaqueValue-typed
attribute value.

(Internal change: 2251717)