Fix broken frame selection functionality in usdview's prim view
Change 2247506 added "Frame Selection" as a QAction using the
"F" key as a shortcut. This caused Qt to bypass the special
keypress event handling in appEventFilter that dispatches
the "F" keypress to the widget with focus and invoke the
camera framing action instead.

The fix modifies appEventFilter to look for the ShortcutOverride
event that Qt provides specifically for cases like this, where
a client wants to override shortcut handling.

This change also fixes a long-standing bug where trying to
frame the selection in the prim view would fail due to a
"NotImplementedError" exception in PySide2 < 5.15. This
was due to QTreeViewItem being compared against None using
__eq__ instead of "is", which was broken in these versions
of PySide2. See https://bugreports.qt.io/browse/PYSIDE-74
for more details.

(Internal change: 2252116)