Fix usdfixbrokenpixarschemas usage when usdz contains a usdc file.
Previous SdfLayer object corresponding to the usdc file was still "in scope" and valid,
while __exit__ of the UsdzAssetIterator was trying to remove the directory containing
this usdc. And since usdc files are memory mapped, windows/nfs would complain
of the file in-use and shutil.rmtree was throwing an exception.

Now since the SdfLayer object is scoped out of the UsdzAssetIterator's with statement,
the object is destroyed before the context manager's __exit__ tries to delete the temp
extracted dir.

Reverted the test usdz to use usdc.

(Internal change: 2252159)