[hd] Fixes a bug in the flattening of "visibility" and "purpose".
Previously, if a prim had the "visibility" key, the flattening code
would not inherit "visibility" from the parent prim.  This is fine if
there is data in the container, but in some cases, we may have the
"visibility" key but no data.  That would result in the flattening logic
getting the default visibility rather than an inherited one.

The same fix is applied for "purpose".

(Internal change: 2252406)