UsdImagingStageSceneIndex Point instancing support:
Replacing code to compute per-instance transform by code that is just forwarding the USD values for translate, scale and rotate.

Note that the transform of a piece of geometry in a prototype is supposed to be relative to the prototype's parent. To achieve this, the instanced by scene index is resetting the xform stack on the prototype root.

The primvars driving the instancer need to have instance interpolation. To achieve this, adding an interpolation token to custom primvar mappings of the primvars data source.

(Internal change: 2252630)