[hdSt] Fix index computation and improve validation
Note the following change in imaging behavior.
BEFORE:
Nonperiodic curves with a vertex count < 4 were not rendered (because the curve segment count was < 0).
Pinned curves with a curve vertex count <= 2 were not rendered correctly due to OOB indexing.

AFTER:
Curves with a vertex count < 2 are not rendered. Nonperiodic curves with a vertex count < 4 are rendered by repeating the last vertex.

(Internal change: 2252710)