Adding the concept of schema family and version to USD schemas.
The family and version of a schema is now parsed from the schema's
identifier and is included in its SchemaInfo. A schema's version is
indicated by a version suffix consisting of a single underscore followed
by a positive integer version number. The schema's family is the
identifier before this suffix. Version 0 of each schema family is the
family name itself, i.e. the family name without a version suffix is
version 0.

The schema registry now has an overload of FindSchemaInfo that takes a
family and version as well as new functions for getting info for all
schemas in a family optionally filtered by a version predicate. It also
has new standalone functions for parsing family and version from
a hypothetical schema identifier, for creating an indentifier from a
hypothetical family and version, and for determining whether a give
family or identifier token is allowed.

(Internal change: 2256429)
(Internal change: 2256558)