Fix a bug in HdDataSourceLegacyPrim introduced in change 2255856
This fixes a bug where binding name was incorrectly reported as
"instanceName:binding" instead of just the "instanceName" which is what
hydra will be looking for.

For backward compatible "coorsSys:instanceName" will still be respected.

This change precedes a change which is introducing CoordSysAPIAdapter
which also looks for the "instanceName" when instancing the datasource
and when querying the corresponding datasource locator.

(Internal change: 2256809)