Fix bug where adding an empty sublayer to a layer stack could cause errors and incorrect results with value clips.
Usd_Clip and Usd_ClipSet previously stored the index of
the layer in the layer stack where the clip set was
authored. When a non-empty sublayer was added to the
layer stack, UsdStage would end up resync'ing prims,
which would ultimately recompute the clip sets and
update that index. However, when an empty sublayer
was added, this resync would not occur and the index
would be invalidated if the sublayer was added before
that position in the layer stack.

Once in this state, UsdStage would incorrectly identify
the layer where the value clips were authored. This could
lead to issues including errors opening value clips
specified via relative asset paths and incorrect results
when looking at an attribute's property stack.

This fix modifies the objects to store the handle of
the layer where the clip set was authored instead of the
index, avoiding the need to update that index on any
sublayer changes.

Fixes #2014

(Internal change: 2256833)