First pass at Hdprman mesh lights, with updates to supporting pieces.
This all allows us to accommodate the Prman requirement that
a separate geometry prototype be provided for mesh lights.

In gprim / gprimbase / mesh:

-Previously, when hdPrman created a mesh, it always created both a
prototype and an instance. This change adds a method to
specify that no instance should be created. A "protoype only" mesh
will provide the prototype geometry for a mesh light.

-This change also adds a method for a mesh to provide its associated
prototype id(s).

In renderDelegate:

-When creating the mesh (rprim) for a mesh light, set its 'isMeshLight'
flag to True (which is what's returned by the _PrototypeOnly method).

In meshLightResolvingSceneIndex:

-We now generate three Hydra prims from the mesh light source prim
(i.e., the one with the Usd API):

1) __meshLight_light - The "light".

2) __meshLight_sourceMesh - The "prototype only" source mesh.

3) __meshLight_blackMaterial - A PxrBlack bxdf, which is bound to both
__meshLight_sourceMesh and __meshLight_light.

There are also updates to correctly handle mesh light "glow" material
duplication and wiring.

In meshLightResolvingSceneIndex / meshLightResolvingSceneIndexPlugin:

-Comments updated.

(Internal change: 2257624)
(Internal change: 2257642)