Adding version aware schema type queries to UsdPrim.
For typed schema queries we now have IsInFamily and
GetVersionIfIsInFamily which will check if the prim answers
IsA to any of a subset of schemas in a determined family.
For API schema queries we now have HasAPIInFamily and
GetVersionHasAPIInFamily which will check if the prim answers
HasAPI to any of a subset of schemas in a determined family.

(Internal change: 2257759)