[HdSt] GPU view frustum culling for pipelineDrawBatch
Implementation of GPU culling using compute shaders in the
PipelineDrawBatch. The vertex-shader transform feedback
approach that was used previously in the IndirectDrawBatch
has not been supported in HGI since it would require multiple
command buffers for synchronisation of the stages.

The compute shader approach does not require a separate
reset pass for instance culling since it can expand out
the instances in a loop within the compute shader and
populate the drawCommand buffers directly.

Since the ICB changes have been merged into dev, this
PR has been rebased on top of this. ICBs need to run
in a separate batch, so they have been moved to the
new Hd_DrawBatch::EncodeDraw method.

Contribution: Jon Creighton

Fixes #2053

(Internal change: 2258394)