When composing a UsdPrimDefinition we now disallow multiple versions of the same API schema family to be applied at the same time.
While composing built-in and auto-apply API schemas into the static
registered USD schema prim definitions, if adding an API schema would
cause a different version of a stronger already composed API schema to
be added, the API schema will be skipped and a warning reported. This
includes both if the API schema itself would cause the version and if
any of the API schema's built-ins that would be added with it would
cause the version conflict.

When composing authored API schemas into a USD prim's prim definition,
the same version conflict checking applies as for built-in API schemas,
but we also check that the authored API schema won't cause any version
conflicts with any API schemas that are included in the prim type's
typed schema definition (even the though the typed schema is technically
weaker, though that should be changing in the future).

In order to facilitate this new feature, this change also includes a
significant refactor to the code that composes the full
UsdPrimDefinitions for API schemas. Previously, each API schema's prim
definition was composed top down in place, in multiple steps, via a
depth first traversal through the entire tree of included API schemas.
This refactor now composes the included API schemas bottom up, which
simplifies the algorithm and allows us to check if an included API
schema and any of its built-ins would cause a version conflict before
composing it.

(Internal change: 2259141)