[arch] Fix error in GetEnvironmentVariable result handling
GetEnvironmentVariable returns result > 0 on success and 0 on error.
If result is 0 we should check GetLastError for actual error code.
In case error code is ERROR_ENVVAR_NOT_FOUND variable is missed.

So if existing env variable should not be overwritten we check
either it present or error occured (but not variable missed error).