[UsdImagingGL] Fixed unitTestGLDrawing arg handling
Updated the argument handling for the "-shading" and "-cullStyle"
options to initialize defaults and only report warnings for unhandled
modes.

Without this fix we get spurious warnings when running through any
of the current unit tests which use the defaults for these arguments.
Also, fixed the diagnostic text to avoid potential confusion
with UsdGeomModelAPI drawMode.

See #2097

(Internal change: 2259725)