Fix bug where Creation of HgiGLSampler will always set GL_TEXTURE_MAX_ANISOTROPY_EXT to 16.
In the constructor of HgiGLSampler, we will check if the minFilter or magFilter is nearest. If yes, we will not set GL_TEXTURE_MAX_ANISOTROPY_EXT.

When GL_TEXTURE_MAX_ANISOTROPY_EXT is set, it will ignore the minFilter and magFilter on some platform. This means that we will never create a nearest sampler.

Formatting