Add support for Python 3.10 to usd-core PyPI package
This change adds 3.10 to the PyPI build matrix on
Azure. Doing so required manually installing CMake 3.24
on top of the manylinux2014 docker image we currently
use, since 3.24+ is required for Python 3.10. Attempting
to use a newer docker image caused crashes when using
the generated wheels for some unknown reason.

(Internal change: 2260771)