[hd, hdPrman, usdImaging] Rework parameters for HdRenderSettings
- Improve granularity of invalidation for the HdRenderSettings Bprim by replacing the singular HdRenderSettingsParams struct (that mapped to the DirtyParams bit) with separate settings and render products parameters mapped to DirtySettings and DirtyRenderProducts bits.

This change requires minor updates to the adapter and emulation layers, and reduces the work done in the adapter. Previously, the resolved "render spec" was computed which involved visiting all the products & vars pointed to by the render settings prim. Note that Hydra 1.0 doesn't support per-settings-attribute invalidation, so we still compute and return a dictionary with _all_ the gathered settings attributes. This change also improves the dirty bits translation in backend emulation (where scene index locators are mapped back to dirty bits).

(Internal change: 2262583)