usdviewq: do not try to fetch properties for UsdCollectionAPI paths when selecting target objects
Since collections can target other collections in UsdCollectionAPI
relationships, it's possible that the target of a relationship is a
UsdCollectionAPI path and not an actual property path. In these cases, we
detect that the path is a UsdCollectionAPI path and avoid trying to get a
property with the collection name, since no such property will exist.

This avoids a warning of the following form when the "Select Target Path"
action is used from the property view context menu in usdview on a relationship
target that is a UsdCollectionAPI path:

Traceback (most recent call last):
  File "C:\USD_inst\lib\python\pxr\Usdviewq\attributeViewContextMenu.py", line 207, in RunCommand
    _selectPrimsAndProps(self._dataModel, paths)
  File "C:\USD_inst\lib\python\pxr\Usdviewq\attributeViewContextMenu.py", line 81, in _selectPrimsAndProps
    raise PropertyNotFoundException(path)
pxr.Usdviewq.common.PropertyNotFoundException: Property not found at path in stage: /CollectionTest/Geom/Shapes.collection:allShapes