Python crashed. FATAL ERROR: Failed axiom: Py_IsInitialized()
PXR_PY_UNDEFINED_DYNAMIC_LOOKUP might be explicitly set when
packaging wheels, or when cross compiling to a Python environment
that is not the current interpreter environment.
If it was not explicitly set to ON or OFF, then determine whether
Python was statically linked to its runtime library by fetching the
sysconfig variable LDLIBRARY, and set the variable accordingly.
If the variable does not exist, PXR_PY_UNDEFINED_DYNAMIC_LOOKUP will
default to OFF. On Windows, LDLIBRARY does not exist, as the default
will always be OFF.

Fixes #1620

(Internal change: 2263328)