build_usd.py will download zip files instead of tarfiles.
This (old) vulnerability in Python's tarfile module made some headlines
late in 2022: https://nvd.nist.gov/vuln/detail/CVE-2007-4559
This change removes use of the tarfile module in favor of downloading zip
files instead. In all cases except the jpeg libraries, the contents of the
zip files have been verified to be the same; the jpeg libraries have crlf
line endings. gcc and clang are fine with this however, the only consequence
will be that if debugging in gdb, line numbers may be off within the jpeg
libraries; but debugging into jpeg is not something a USD developer is
likely to be doing.

It also fixes an error in the zip file-filtering routine, exposed by switching
the boost download to zip.

(Internal change: 2263346)