[hdMtlx] Fix for MaterialX materials with the name 'material'
When we import all the MaterialX libraries, MaterialX adds a
child to the document with the name 'material' because of
the typedef defined in stdlib_defs.mtlx. So when we attempt
to add a material of the same name to the document we get
an error. This change uses of the createValidChildName helper
function to avoid this error.

Fixes #2076

(Internal change: 2263534)