[hd, usdImaging] Add support for transporting render products and vars
- Update HdRenderSettings Bprim to sync render products.
- Add render product and var schemas.
- Update scene index emulation.
- Update render settings adapter to return the flattened render products in Get(..).

Note that a flattened/aggregated representation is used for the render products generated by a render settings prim and for the render vars that comprise a render product for simplicity sake. Specifically, this avoids introducing additional 1.0 Hydra prim types for products and vars and working around existing limitations in the (1.0) Sync algorithm.

(Internal change: 2263557)