[usdImaging] Guard against missing instancer data in GetInstanceCategories
The scene delegate passes requests for point instancer data to the native instancing adapter when the point instancer is part of the prototype of a native instance. We rely on this adapter hijacking to support selection of all instances of a prototype, regardless of instancing type and hierarchy.

Changes introduced in 2264225 expected the native instancing adapter's GetInstanceCategories to only ever be called for native instancers. When confronted with a point instancer, a fatal crash occurred.

Rather than remove the need for adapter hijacking (since that's a major undertaking), or detecting and deferring to the proper adapter (since we may not have all the information needed to do that), we will instead simply gate retrieval of instance categories behind a safety check. Such a check was present prior to 2264225; this change simply restores that check. The resulting output of GetInstanceCategories will be the same as if we had deferred to the point instancer adapter: an empty vector.

(Internal change: 2264956)