Small cleanups for token-related issues in API schemas
- Remove unused prim alias token from _schemaTokens

This token had previously been used by the Apply function for
applied API schemas but is no longer needed thanks to recent
changes to how that function is implemented.

For non-applied and single-apply API schemas, this token was
the only member of the _schemaTokens collection, so the entire
collection was removed. Doing so fixes unused-variable warnings
emitted by clang.

- Remove unused usd/tokens.h include

This was also unnecessary due to changes in the implementation
of Apply.

(Internal change: 2265666)