pcp:  Internal references within variants now work.  They also now offer namespace visibility to the rest of the layer stack.
Background

Resolution of variant selections is conceptually simple.
We defer variant evaluation until no other unevaluated arcs
remain, to ensure that maximum opinion information is available to
decide the variant.  At that point, we resolve variant sets
in strength order.  For each, the strongest variant selection
opinion prevails.

Implementation concerns complicate this picture in practice.
For one, composition arcs to sub-root namespace locations require
consideration of ancestral opinions -- i.e., those on namespace
parents of the target of the arc, since these may affect composition
of the child.

For another, composition is a recursive algorithm that builds
sub-graphs before integrating them into the caller's graph.
This complicates strength-order traversals -- such as those
used to decide variant selections -- while the structure
is being still built.

To implement variant resolution given these considerations, we
maintain an explicit stack frame of recursive indexer calls, and
use that to guide strength-order traversal across the subgraphs
that will eventually form the final prim index graph.  This is
also used map namespace, which is crucial to finding the
right opinions to index.

In the case of internal reference arcs, i.e. those within
the same layerstack, discovery of ancestral variant
selections was failing due to the namespace-encapsulating
nature of references.

Specifically, references encapsulate their target namespace and
prohibit visibility to anything outside it.  As a consequence,
when recursing up ancestral prim graphs of a sub-root internal
reference, there was no longer any valid namespace mapping.
This meant there was no way to discover the ancestral variant
selection in the calling indexer stack frame.

Stepping back, we reconsider our policy.  Reference arcs by
design have restrictive namespace mapping in order to provide
reliable encapsulation behavior.  However, the addition of
sub-root and internal references brings reason to reconsider this.
An internal reference is not a case of cross-asset reuse, which
is where encapsulation is so important for predictable behavior;
instead, it is a case of re-use internal to the implemention of
an asset.  Therefore, as in the case of a local inherit arc,
it make sense to provide a root identity mapping (i.e. </> :
</i>) so that paths outside the source & target of the internal
reference are available.

Doing this makes the ancestral variant selection
visible across the internal sub-root reference, making
the expected behavior "just work" as a consequence.

(Internal change: 2265786)