Fix a race condition when ArFilesystemWritableAsset creates directories
When an asset is opened for writing, ArFilesystemWritableAsset will
automatically create directories in the asset's path. Pass true to
the existsOk argument of TfMakeDirs to allow the asset to be opened
even if the directory is created by another thread/process at the
same time ArFilesystemWritableAsset is trying to create it.

Comes with a test that verifies the clean simultaneous opening of
assets that require directory hierarchies to be created.

(Internal change: 2266556)
(Internal change: 2266721)