[hdPrman] Add a pathway in the HdPrman test harness to use the RenderSettings prim instead of the UsdRenderSpec
In order to trigger using a RenderSettings prim instead of the RenderSpec,
the 'useRenderSettings' test option needs to be used. Without this option
the test will revert back to the fallback bdhavior of using the UsdRenderSpec.

Note that there are two ways for the test to get the authored Render Settings
prim. Either by specifying the path in the stage metadata, or by giving the
path with the 'settings' test option. If there is an authored Render Settings
prim but the path is not specified either in the metadata or the test option,
the test will use the fallback RenderSpec.

Note that the HdRenderSettingsMap is still being used to get the settings
to the RenderPass. What is changing is that the information is coming
from the Hydra RenderSettings Prim instead of the UsdRenderSpec.

(Internal change: 2268136)