The UsdSchemaRegistry no longer combines all the plugin generatedSchema files into a single large schematics layer and instead just allows the UsdPrimDefinitions it builds to map its properties directly to the generateSchema layer that holds its prim spec. This is a performance improvement in the initialization of the schema registry and opens up potential further improvement to the generatedSchemas themselves.
Refactored some of the prim definition building implementation, moving
it from UsdSchemaRegistry into UsdPrimDefinition with the goal of having
all the code that handles the _LayerAndPath structure live in the
UsdPrimDefinition code itself.

Upgraded _LayerAndPath from a std::pair to a struct with helper accessor
members for the now common uses of accessing data from the layer and
path.

(Internal change: 2268201)