[usdImaging] Improve cards performance in storm
The changeover to UsdPreviewSurface for cards attempted to improve visual quality of texture-based cutouts, but this led to performance issues in Storm. This change returns to using opacityThreshold to signal that transparent parts of the card texture should be cut out. This restores performance in Storm at the cost of cutout edge artifacts in both Storm and Prman. Baselines are updated in multiple components to account for the new artifacts.

Also includes a fix to set wrapS and wrapT on the texture reader node correctly to reduce outside-edge artifacts around cards, separately from cutout presence.

(Internal change: 2268365)