[usdz, ar] Support ArAsset / ArWritableAsset for USDZ
[usdz] USDZ zip files would always force the use of file system paths.
For asset management systems that don't directly rely on the file system this could be inconvenient. It would require downloading content to temp files, writing to a temp file, then uploading the temp zip file. With Ar 2.0, we can now use OpenAsset / OpenAssetForWrite so the underlying ArResolver can manage this directly. This change adds support so we can write directly to an ArWritableAsset to abstract the file system.
[ar] Required a change to ArFilesystemWritableAsset to discard the TfSafeOutput file if Close() was not explicitly called.