[HdSt] Simplified primitiveCoord plumbing
To index into primitive data buffers for the current primitive
we need to consider two offsets: primitiveCoord which is the
offset for the current draw invocation and PrimitiveID which is
the offset for the current primitive within the draw invocation.

Previously, we applied (and unapplied) these offsets while plumbing
the primitveCoord between shader stages. Now, this is done within
the GetPrimitiveIndex() accessor method.

(Internal change: 2268600)