This change was inspired by the work done from github pull request #2133 and contains the following items:
Creates a new UsdShadeUdimUtils class which has methods for working with Udim files.  This includes detecting the presence of a UDIM identifier, reoslving UDIM paths, and subsituting indices back into an identifier string. The core logic for these methods was lifted from UsdImaging's materialParamUtils

Updates UsdUtils dependencies resolution logic to use these new methods to ensure UDIM tiles are are correctly included as dependencies.

Updates UsdImaging's material parsm to make use of this new method

As a result of this change, usdUtils now has a dependency on usdShade

Thanks to @lbiasco for the foundations of this change!

Fixes #2133

(Internal change: 2268797)